refactor: add `types`
